Add files via upload
Rough UI Sketch 01

Hi Guys, here is the start of some super rough UI thoughts for Eye Forge.

Please don't take these as final as they are extremely rough, but hopefully just give a starting point of what things could be.
— I've tried to keep UI elements down to a minimum and as simple as possible.
— I'm not sure on how accurate gaze selection is according to space/pixels on screen.  
— There is a possibility that some of these elements may be too intricate, but let me know if so.

Overview:

All UI elements would be grey until gazed/selected and would change colour (and stay that colour) accordingly.

TOP LEFT
Simple reset button

BOTTOM LEFT
This is the trickiest element, and I'm not sure if this is right just yet, but I'm keen for people to test it out. Ideally, this would be more of a Spirograph-style dial, but that isn't possible based on a gaze system really. It consists of a 'pie' cut into eight elements that go into positive/negative angles and importantly, a 0 or no-angle control in the middle. Again... I'm not sure this is the best way to visualise this controller... I've some other thoughts but I need time to think on them.

TOP RIGHT
Speed controller. Almost like a throttle. +4 and -4 from top to bottom. A stop function in the middle - feels super important. Currently, you have that function as '0' but I think it is more effective as a stop.

DIRECTION
You haven't got this at the moment, and it may require further coding, but for me, it feels like it could be a simple and important control that would help you draw squares and other shapes incredibly easily.

See what you think to the screenshot. I'm not sure it is all possible, but if you want to test anything out, let me know how best to export the elements for you!

Andy